7/11 ヨシダ_コンテンツ詳細一覧、コンテンツ詳細削除作成(未テスト)

+++ VoteSystem/WebContent/manager.jsp
@@ -1,5 +1,10 @@
+<%@page import="model.ContentsBean"%>
+<%@page import="java.util.ArrayList"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
 	pageEncoding="UTF-8"%>
+<%
+	ArrayList<ContentsBean> contents = (ArrayList<ContentsBean>)session.getAttribute("contents");
+%>
 <!DOCTYPE html>
 <html>
 <head>
@@ -56,53 +61,19 @@
 				</div>
 				<h1>コンテンツ確認</h1>
 				<div class="row">
+					<% for(int i = 0; i < contents.size(); i++){ %>
 					<div class="col-xl-4 col-lg-4 col-md-6 col-sm-12 col-12 textcenter">
-						<img class="img-responsive" src="img/no_image.png" width="300" height="200" alt="コンテンツ画像">
+						<img class="img-responsive" src="${pageContext.request.contextPath}/GetContentsPicture?id=<%= contents.get(i).getContentsID() %>" width="300" height="200" alt="コンテンツ画像">
 					</div>
 					<div class="col-xl-2 col-lg-2 col-md-6 col-sm-12 col-12 listmargin">
 						<div class="list-group">
 							<a href="manager_contentsUpdate.jsp" class="list-group-item list-group-item-action">情報更新</a>
-							<a href="manager_contentsData.jsp" class="list-group-item list-group-item-action">参加者</a>
+							<a href="GetAllContentsdata?id=<%= contents.get(i).getContentsID() %>" class="list-group-item list-group-item-action">参加者</a>
 							<a href="manager_contentsStatus.jsp" class="list-group-item list-group-item-action">状況確認</a>
-							<a href="manager_contentsDelete.jsp" class="list-group-item list-group-item-action" onclick="return confirm('退会してもいいですか？')">削除</a>
+							<a href="manager_contentsDelete.jsp" class="list-group-item list-group-item-action" onclick="return confirm('削除してもいいですか？')">削除</a>
 						</div>
 					</div>
-
-					<div class="col-xl-4 col-lg-4 col-md-6 col-sm-12 col-12 textcenter">
-						<img class="img-responsive" src="img/no_image.png" width="300" height="200" alt="コンテンツ画像">
-					</div>
-					<div class="col-xl-2 col-lg-2 col-md-6 col-sm-12 col-12 listmargin">
-						<div class="list-group">
-							<a href="manager_contentsUpdate.jsp" class="list-group-item list-group-item-action">情報更新</a>
-							<a href="manager_contentsData.jsp" class="list-group-item list-group-item-action">参加者</a>
-							<a href="manager_contentsStatus.jsp" class="list-group-item list-group-item-action">状況確認</a>
-							<a href="manager_contentsDelete.jsp" class="list-group-item list-group-item-action">削除</a>
-						</div>
-					</div>
-					
-					<div class="col-xl-4 col-lg-4 col-md-6 col-sm-12 col-12 textcenter">
-						<img class="img-responsive" src="img/no_image.png" width="300" height="200" alt="コンテンツ画像">
-					</div>
-					<div class="col-xl-2 col-lg-2 col-md-6 col-sm-12 col-12 listmargin">
-						<div class="list-group">
-							<a href="manager_contentsUpdate.jsp" class="list-group-item list-group-item-action">情報更新</a>
-							<a href="manager_contentsData.jsp" class="list-group-item list-group-item-action">参加者</a>
-							<a href="manager_contentsStatus.jsp" class="list-group-item list-group-item-action">状況確認</a>
-							<a href="manager_contentsDelete.jsp" class="list-group-item list-group-item-action">削除</a>
-						</div>
-					</div>
-
-					<div class="col-xl-4 col-lg-4 col-md-6 col-sm-12 col-12 textcenter">
-						<img class="img-responsive" src="img/no_image.png" width="300" height="200" alt="コンテンツ画像">
-					</div>
-					<div class="col-xl-2 col-lg-2 col-md-6 col-sm-12 col-12 listmargin">
-						<div class="list-group">
-							<a href="manager_contentsUpdate.jsp" class="list-group-item list-group-item-action">情報更新</a>
-							<a href="manager_contentsData.jsp" class="list-group-item list-group-item-action">参加者</a>
-							<a href="manager_contentsStatus.jsp" class="list-group-item list-group-item-action">状況確認</a>
-							<a href="manager_contentsDelete.jsp" class="list-group-item list-group-item-action">削除</a>
-						</div>
-					</div>
+					<% } %>
 				</div>
 			</div>
 		</div>
